fix: render을 위한 함수들을 view로 이동한다.
