+++ src/controller/productManage/productManageController.ts
@@ -1,6 +1,6 @@
 import ProductManageModel from '../../model/productManageModel'
 import ProductStore from '../../store/ProductStore'
-import { ADD_PRODUCT, RENDER_PRODUCT } from '../../utils/constants/action'
+import { ADD_PRODUCT } from '../../utils/constants/action'
 import { ProductManageReucerAction } from './productManageAction'
 
 export default class ProductManageController {
@@ -14,10 +14,6 @@ export default class ProductManageController {
     switch (action.type) {
       case ADD_PRODUCT:
         this.#model.addProduct(action.payload)
-        break
-
-      case RENDER_PRODUCT:
-        this.#model.renderProduct()
         break
     }
   }
