+++ src/utils/constants/action.ts
@@ -1,4 +1,3 @@
 const ADD_PRODUCT = 'ADD_PRODUCT'
-const RENDER_PRODUCT = 'RENDER_PRODUCT'
 
-export { ADD_PRODUCT, RENDER_PRODUCT }
+export { ADD_PRODUCT }
